re-organized where the special case serialization lives
